whisperApi/requirements.txt
flask
openai-whisper
torch
